+++ lib/src/base32.dart
@@ -76,6 +76,5 @@ Uint8List fromBase32(
   bool padding = true,
 }) {
   codec ??= _codecFromParameters(padding: padding);
-  var out = codec.decoder.convert(input.codeUnits);
-  return Uint8List.fromList(out);
+  return codec.decoder.convert(input.codeUnits);
 }